Shyauna Hood
This week I could not get any of my codes to run and I am not sure why. I also just can't figure out the math parts of the code. How to calculate formulas and such so the Fahrenheit was really difficult for me. All of my codes kept giving me the same error and I tried having classmates help me, but it didn't work. I also opened your ink in my program and it would not run either so I'm not sure if it's just my program or what.